fix(PUCSchema): Refactor schema name for PUC
